new interest/time estimators

+++ DiversificationInit.cpp
@@ -32,7 +32,7 @@ namespace cplex_tap {
                 }
                 S.emplace_back(tmp);
             }
-            vector<int> time = JVMAdapter::getTime(baseSet, pricingIST);
+            vector<double> time = JVMAdapter::getTime(baseSet, pricingIST);
             vector<double> interest = JVMAdapter::getInterest(baseSet, pricingIST);
 
             /*
